requirements.txt
flask>= 2.3.0
requests>= 2.28.0
pandas>= 2.0.0
numpy>= 1.24.0
pillow>= 9.0.0
transformers>= 4.35.0
# torch>= 2.1.0  For machine learning tasks if you need cpu support
google-generativeai
opencv-python
